+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added 
+
+- Add `icon` component.
+
+### Changed
+
+- Update tools.
+
+### Fixed
+
+- Typo in `Alert` component.
+
 ## 0.3.0 - 2025-12-19
 
 ### Added